Finally

+++ Kanban-app/backend/src/usertasks/delete-a-task/controllers/delete-task.controller.ts
@@ -17,24 +17,14 @@ export class DeleteTaskControl {
     public async deleteTaskControl(@Query() task:GeneralDetailsDTO):Promise<string> {
         let tag:string;
         let message:string;
-        const unsafeMessage:string | void = await this.allCheckService.checkForAll(task.boardName,task.groupName,'',true,task.index)
-        if (unsafeMessage === 'board not found' || 'group not found' || 'task not found') {
-            tag = 'UNSAFE'
-        }else if (unsafeMessage === 'board not found') {
-            message = `The board,'${task.boardName}' doesnt exist to perform the delete operation`
-            return `${tag}:${message}`
-        }else if (unsafeMessage === 'group not found') {
-            message = `The group,'${task.groupName}' doesnt exist to perform the delete operation`
-            return `${tag}:${message}`
-        }else if (unsafeMessage === 'task not found') {
-            message = `The board,'${task.boardName}' doesnt exist to perform the delete operation`
-            return `${tag}:${message}`
+        const unsafeMessage:string | string[] = await this.allCheckService.checkOperationSafety('delete',task.boardName,task.groupName,true,task.index)
+        if (!Array.isArray(unsafeMessage)) {
+            return unsafeMessage 
         }
-
+        let taskTitle = unsafeMessage[1]
         await this.deleteTaskService.deleteTask(task.boardName,task.groupName,task.index);
-        const taskTitle = unsafeMessage
         tag = 'SUCCESSFUL';
         message = `Deleted the task '${taskTitle}' from the group '${task.groupName}' from the board '${task.boardName}'`
-        return `${tag}:${message}`
+        return `${tag}:${message}`;
     }
 }
